+++ crates/nostr-relay-pool/src/lib.rs
@@ -20,6 +20,7 @@ pub mod relay;
 #[doc(hidden)]
 mod shared;
 mod stream;
+pub mod transport;
 
 pub use self::pool::options::RelayPoolOptions;
 pub use self::pool::{Output, RelayPool, RelayPoolNotification};